+++ athena/playground/src/components/view_mode/evaluation_mode/index.tsx
@@ -6,7 +6,7 @@ import { useState } from "react";
 import DefineExperiment from "./define_experiment";
 import ConfigureModules from "./configure_modules";
 import ConductExperiment from "./conduct_experiment";
-import EvaluationManagement from "@/components/view_mode/expert_evaluation/evaluation_management";
+import ComparativeEvaluationStudyManagement from "@/components/view_mode/comparative_evaluation_study/index";
 
 export default function EvaluationMode() {
   const [experiment, setExperiment] = useState<Experiment | undefined>(
@@ -42,7 +42,7 @@ export default function EvaluationMode() {
         </>
       )}
       <h2 className="text-4xl font-bold text-white mb-4">Expert Evaluation</h2>
-      <EvaluationManagement />
+      <ComparativeEvaluationStudyManagement />
     </>
   );
 }
